@@ -1,14 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
+    <PublishAot>true</PublishAot>
   </PropertyGroup>
 
   <PropertyGroup>
-    <Version>1.0.4.1</Version>
+    <InvariantGlobalization>true</InvariantGlobalization>
+    <IlcGenerateCompleteTypeMetadata>false</IlcGenerateCompleteTypeMetadata>
+    <IlcTrimMetadata>true</IlcTrimMetadata>
+    <IlcGenerateStackTraceData>false</IlcGenerateStackTraceData>
+    <IlcOptimizationPreference>Size</IlcOptimizationPreference>
+    <IlcFoldIdenticalMethodBodies>true</IlcFoldIdenticalMethodBodies>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <Version>1.0.4.2</Version>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CryMediaAPI" Version="1.1.2" />
+    <PackageReference Include="CryMediaAPI" Version="1.2.0" />
   </ItemGroup>
 </Project>
